Fix stories / heirarchy
